feat: move helpers to test fixture
